+++ pygsp/filters/itersine.py
@@ -46,7 +46,7 @@ class Itersine(Filter):
         for i in range(1, Nf + 1):
             g.append(lambda x, ind=i: k(x/scale - (ind - overlap/2.)/overlap) / np.sqrt(overlap)*np.sqrt(2))
 
-        super(Itersine, self).__init__(G, g, **kwargs)
-
-        self.mu = np.linspace(0,G.lmax,num=Nf)
+        self.mu = np.linspace(0, G.lmax, num=Nf)
         self.overlap = overlap
+
+        super(Itersine, self).__init__(G, g, **kwargs)