+++ pygsp/learning.py
@@ -1,8 +1,21 @@
 # -*- coding: utf-8 -*-
 
 r"""
-The :mod:`pygsp.learning` module implements some learning functions used
-throughout the package.
+The :mod:`pygsp.learning` module implements functions to solve learning
+problems.
+
+Semi-supervized learning
+========================
+
+Those functions help to solve a semi-supervized learning problem, i.e. a
+problem where only some values of a graph signal are known and the others shall
+be inferred.
+
+.. autosummary::
+
+    regression_tik
+    classification_tik
+
 """
 
 import numpy as np
@@ -88,56 +101,62 @@ def classification_tik_simplex(G, y, M, tau=0.1, **kwargs):
 
 
 def classification_tik(G, y, M, tau=0):
-    ''' Solve a semi-supervised classification problem on the graph.
+    r"""Solve a classification problem on graph via Tikhonov minimization.
 
-        The function first transform y in logits Y. Then it solves
+    The function first transform :math:`y` in logits :math:`Y`. It then solves
 
-            argmin_X   || M X - Y ||_2^2 + tau tr(X^T L X)
+    .. math:: \operatorname*{arg min}_X \| M X - Y \|_2^2 + \tau \ tr(X^T L X)
 
-        if tau > 0 and
+    if :math:`\tau > 0` and
 
-            argmin_X   tr(X^T L X)   s. t.  Y = M X
+    .. math:: \operatorname*{arg min}_X tr(X^T L X) \ \text{ s. t. } \ Y = M X
 
-        otherwise, where X and Y are logits. Eventually, the function return
-        the maximum of the logits.
+    otherwise, where :math:`X` and :math:`Y` are logits. The function return
+    the maximum of the logits.
 
-        Inputs
-        ------
-        G : Graph
-        y : Measurements (numpy array [G.N, :])
-        M : Mask (vector of boolean [G.N,])
-        tau : regularization parameter
+    Parameters
+    ----------
+    G : Graph
+    y : array of length G.N
+        Measurements
+    M : array of boolean, length G.N
+        Masking vector.
+    tau : float
+        Regularization parameter.
 
-        Example
-        -------
-        import numpy as np
-        import pygsp
-        import matplotlib.pyplot as plt
-        pygsp.plotting.BACKEND = 'matplotlib'
+    Examples
+    --------
+    >>> from pygsp import graphs, learning
+    >>> import matplotlib.pyplot as plt
+    >>>
+    >>> G = graphs.Logo()
 
-        G = pygsp.graphs.Logo()
-        idx_g = np.squeeze(G.info['idx_g'])
-        idx_s = np.squeeze(G.info['idx_s'])
-        idx_p = np.squeeze(G.info['idx_p'])
-        sig = np.zeros([G.N])
-        sig[idx_s] = 1
-        sig[idx_p] = 2
+    Create a ground truth signal:
 
-        # Make the input signal
-        M = np.random.uniform(0,1,[G.N])>0.5 # Mask
+    >>> signal = np.zeros(G.N)
+    >>> signal[G.info['idx_s']] = 1
+    >>> signal[G.info['idx_p']] = 2
 
-        measurements = sig.copy()
-        measurements[M==False] = np.nan
+    Construct a measurements signal from a binary mask:
 
-        # Solve the classification problem
-        recovery = pygsp.learning.classification_tik(G, measurements, M, tau=0)
+    >>> rs = np.random.RandomState(42)
+    >>> mask = rs.uniform(0, 1, G.N) > 0.5
+    >>> measurements = signal.copy()
+    >>> measurements[~mask] = np.nan
 
-        # Plotting
-        f, (ax1, ax2, ax3) = plt.subplots(1, 3, sharey=True, figsize=(14,4))
-        G.plot_signal(sig, plot_name = 'Ground truth', ax=ax1)
-        G.plot_signal(measurements, plot_name = 'Measurement', ax=ax2)
-        G.plot_signal(recovery, plot_name = 'Recovery', ax=ax3)
-    '''
+    Solve the classification problem by reconstructing the signal:
+
+    >>> recovery = learning.classification_tik(G, measurements, mask, tau=0)
+
+    Plot the results:
+
+    >>> fig, (ax1, ax2, ax3) = plt.subplots(1, 3, sharey=True, figsize=(10, 3))
+    >>> G.plot_signal(signal, plot_name='Ground truth', ax=ax1)
+    >>> G.plot_signal(measurements, plot_name='Measurements', ax=ax2)
+    >>> G.plot_signal(recovery, plot_name='Recovery', ax=ax3)
+    >>> fig.tight_layout()
+
+    """
 
     def to_logits(x):
         l = np.zeros([len(x), np.max(x)+1])
@@ -151,56 +170,60 @@ def classification_tik(G, y, M, tau=0):
 
 
 def regression_tik(G, y, M, tau=0):
-    ''' Solve a regression problem on the graph
+    r"""Solve a regression problem on graph via Tikhonov minimization.
 
-        If tau > 0:
+    If :math:`\tau > 0`:
 
-            argmin_x   || M x - y ||_2^2 + tau x^T L x
+    .. math:: \operatorname*{arg min}_x \| M x - y \|_2^2 + \tau \ x^T L x
 
-        else:
+    else:
 
-            argmin_x   x^T L x    s. t.  y = M x
+    .. math:: \operatorname*{arg min}_x x^T L x \ \text{ s. t. } \ y = M x
 
-        Inputs
-        ------
-        G : Graph
-        y : Measurements (numpy array [G.N, :])
-        M : Mask (vector of boolean [G.N,])
-        tau : regularization parameter
+    Parameters
+    ----------
+    G : Graph
+    y : array of length G.N
+        Measurements
+    M : array of boolean, length G.N
+        Masking vector.
+    tau : float
+        Regularization parameter.
 
-        Example
-        -------
-        import numpy as np
-        import pygsp
-        import matplotlib.pyplot as plt
-        pygsp.plotting.BACKEND = 'matplotlib'
+    Examples
+    --------
+    >>> from pygsp import graphs, filters, learning
+    >>> import matplotlib.pyplot as plt
+    >>>
+    >>> G = graphs.Sensor(N=100, seed=42)
+    >>> G.estimate_lmax()
 
-        # Create the graph
-        G = pygsp.graphs.Sensor(N=100)
-        G.estimate_lmax()
+    Create a smooth ground truth signal:
 
-        # Create a smooth signal
-        filt = lambda x: 1 / (1+10*x)
-        g_filt = pygsp.filters.Filter(G,filt)
-        sig = g_filt.analyze(np.random.randn(G.N,1))
+    >>> filt = lambda x: 1 / (1 + 10*x)
+    >>> g_filt = filters.Filter(G, filt)
+    >>> rs = np.random.RandomState(42)
+    >>> signal = g_filt.analyze(rs.randn(G.N))
 
-        # Make the input signal
-        M = np.random.uniform(0,1,[G.N])>0.5 # Mask
+    Construct a measurements signal from a binary mask:
 
-        measurements = sig.copy()
-        measurements[M==False] = np.nan
+    >>> mask = rs.uniform(0, 1, G.N) > 0.5
+    >>> measurements = signal.copy()
+    >>> measurements[~mask] = np.nan
 
-        # Reconstructing the signal
-        recovery = pygsp.learning.regression_tik(G, measurements, M, tau=0)
+    Solve the regression problem by reconstructing the signal:
 
-        # Plotting
-        f, (ax1, ax2, ax3) = plt.subplots(1, 3, sharey=True, figsize=(9,3))
-        c = [np.min(sig), np.max(sig[:,0])]
-        G.plot_signal(sig, plot_name = 'Ground truth', ax=ax1, limits=c)
-        G.plot_signal(measurements plot_name = 'Measurement', ax=ax2, limits=c)
-        G.plot_signal(recovery, plot_name = 'Recovery', ax=ax3, limits=c)
+    >>> recovery = learning.regression_tik(G, measurements, mask, tau=0)
 
-    '''
+    Plot the results:
+
+    >>> f, (ax1, ax2, ax3) = plt.subplots(1, 3, sharey=True)
+    >>> c = [signal.min(), signal.max()]
+    >>> G.plot_signal(signal, plot_name='Ground truth', ax=ax1, limits=c)
+    >>> G.plot_signal(measurements, plot_name='Measurements', ax=ax2, limits=c)
+    >>> G.plot_signal(recovery, plot_name='Recovery', ax=ax3, limits=c)
+
+    """
 
     if tau > 0:
         y[M == False] = 0
@@ -238,7 +261,7 @@ def regression_tik(G, y, M, tau=0):
         Luu = G.L[indu, :][:, indu]
         Wul = - G.L[indu, :][:, indl]
         if type(Luu).__module__ == np.__name__:
-            sol_part = np.linalg.solve(Luu, np.matmul(Wul,y[indl]))
+            sol_part = np.linalg.solve(Luu, np.matmul(Wul, y[indl]))
         else:
             sol_part = scipy.sparse.linalg.spsolve(Luu, Wul.dot(y[indl]))